microservices/menu_chat_service/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
google-generativeai>=0.3.0
scikit-learn>=1.2.2
numpy>=1.24.3
requests>=2.28.2
python-dotenv>=1.0.0 